backend/requirements.txt

openai
psycopg2
fastapi
fastapi-login
Flask
Flask-SQLAlchemy
SQLAlchemy
uvicorn[standard]
python-dotenv
cloudinary
python-multipart
langchain
langdetect
elasticsearch==8.15.1
aiohttp>=3.8.0
pydantic